Models predicting customer purchase
This model was build as a part of an assignment/ project in BUS5PA. The aim of the model was to predict whether the customers will purchase from the new clothing line introduced by the client or not. Further, it was used to identify the optimal target audience for marketing.