Model for Churn Analysis
The model was build as an assignment/project for BUS5CA. The aim of the model was to predict whether the customer will churn or not. As, the data set was largely imbalanced, oversampling was done for the minority class( non-churners) and several models were created and analysed. Moreover, binary encoding and standardisation was also done to deal with categorical and varied numerical variables respectively.