Customer Segmentation ( Demographic Segments)
This assignment/project was done under BUS5CA. The primary aim of this was to perform cluster analysis to understand the different characteristics of the company's customer base. Clustering was done on segregating customer data into demographic and behavioural variables. Thus, two segments were made based upon the demographic and behavioural variables.